+++ Game Engine/RQMob.h
@@ -1,14 +1,25 @@
 #import <Foundation/Foundation.h>
 #import <CoreData/CoreData.h>
 
+enum {
+    RQElementalTypeNone = 0,
+    RQElementalTypeFire = 1,
+    RQElementalTypeWater = 2,
+    RQElementalTypeEarth = 3,
+    RQElementalTypeAir = 4,
+};
+typedef NSUInteger RQElementalType;
+
 @interface RQMob : NSManagedObject 
 {
 
 }
 
 @property (nonatomic, retain) NSString *name;
+@property (nonatomic, retain) NSString *spriteImageName;
 @property (nonatomic, retain) NSNumber *currentHPAsNumber;
 @property (nonatomic, retain) NSNumber *levelAsNumber;
+@property (nonatomic, retain) NSNumber *typeAsNumber;
 @property (nonatomic, retain) NSNumber *experiencePointsAsNumber;
 @property (nonatomic, retain) NSNumber *staminaAsNumber;
 @property (nonatomic, retain) NSNumber *staminaRegenRateAsNumber;
@@ -17,12 +28,13 @@
 @property (readwrite, assign) NSInteger currentHP;
 @property (readonly, assign) NSInteger maxHP;
 @property (readwrite, assign) NSInteger level;
+@property (readwrite, assign) RQElementalType type;
 @property (readwrite, assign) NSInteger experiencePoints;
 @property (readwrite, assign) float stamina;
 @property (readwrite, assign) float staminaRegenRate;
 @property (readwrite, assign) NSInteger secondsLeftOfShields;
 
-- (NSInteger)randomAttackValueAgainstMob:(RQMob *)mob;
+- (NSInteger)randomAttackValueAgainstMob:(RQMob *)mob withWeaponOfType:(RQElementalType)weaponType;
 
 + (NSInteger)experinceNeededToLevelFromLevel:(NSInteger)level;
 - (NSInteger)experiencePointsWorth;
@@ -31,5 +43,7 @@
 - (NSInteger)baseAttackPower;
 
 - (NSArray *)weapons;
+- (RQElementalType)weakToType;
+- (RQElementalType)strongToType;
 
 @end