+++ Game Engine/RQMob.m
@@ -15,8 +15,10 @@
 }
 
 @dynamic name;
+@dynamic spriteImageName;
 @dynamic currentHPAsNumber;
 @dynamic levelAsNumber;
+@dynamic typeAsNumber;
 @dynamic experiencePointsAsNumber;
 @dynamic staminaAsNumber;
 @dynamic staminaRegenRateAsNumber;
@@ -46,6 +48,29 @@
 
 - (void)setLevel:(NSInteger)value {
 	[self setLevelAsNumber:[NSNumber numberWithInteger:value]];
+}
+
+- (RQElementalType)type {
+	RQElementalType answer;
+	switch ([self.typeAsNumber integerValue]) {
+		case RQElementalTypeFire:
+			answer = RQElementalTypeFire;
+			break;
+		case RQElementalTypeWater:
+			answer = RQElementalTypeWater;
+			break;
+		case RQElementalTypeEarth:
+			answer = RQElementalTypeWater;
+			break;
+		case RQElementalTypeAir:
+			answer = RQElementalTypeAir;
+			break;
+	}
+	return answer;
+}
+
+- (void)setType:(RQElementalType)value {
+	[self setTypeAsNumber:[NSNumber numberWithInteger:value]];
 }
 
 - (NSInteger)experiencePoints {
@@ -86,15 +111,20 @@
 	[self setSecondsLeftOfShieldsAsNumber:[NSNumber numberWithInteger:value]];
 }
 
-- (NSInteger)randomAttackValueAgainstMob:(RQMob *)mob
+- (NSInteger)randomAttackValueAgainstMob:(RQMob *)mob withWeaponOfType:(RQElementalType)weaponType
 {
 	// baseAttack power * rand(.5 - 1.5) * stamina effect .. when stam is full it is 1.0
 	float randomAttackPower = self.baseAttackPower * ((((rand()%11)+5.0)/10.0));
 	float staminaEffect = pow(self.stamina, 3.0);
-	NSInteger result = round(randomAttackPower * staminaEffect);
-	//NSLog(@"randomAttackPower %f * staminaEffect %f = %d", randomAttackPower, staminaEffect, result);
+	float weaponEffect = 1.0;
+	if ([mob weakToType] == weaponType) {
+		weaponEffect = 2.0;
+	} else if ([mob strongToType] == weaponType) {
+		weaponEffect = 0.5;
+	}
+	NSInteger result = lroundf(randomAttackPower * staminaEffect * weaponEffect);
+	NSLog(@"randomAttackPower %f * staminaEffect %f * weaponEffect %f = %d", randomAttackPower, staminaEffect, weaponEffect, result);
 	return result;
-	// TODO: double the attack value is the mob is weak to the weapon being used
 }
 
 + (NSInteger)experinceNeededToLevelFromLevel:(NSInteger)level
@@ -148,11 +178,51 @@
 	// Fire has dominance over Earth.
 	// Earth has dominance over Air.
 	return [NSArray arrayWithObjects:
-			[NSDictionary dictionaryWithObjectsAndKeys:@"fire", @"type", @"earth", @"strongTo", @"water", @"weakTo", [UIColor redColor], @"color", nil],
-			[NSDictionary dictionaryWithObjectsAndKeys:@"water", @"type", @"fire", @"strongTo", @"wind", @"weakTo", [UIColor blueColor], @"color", nil],
-			[NSDictionary dictionaryWithObjectsAndKeys:@"earth", @"type", @"wind", @"strongTo", @"fire", @"weakTo", [UIColor brownColor], @"color", nil],
-			[NSDictionary dictionaryWithObjectsAndKeys:@"air", @"type", @"water", @"strongTo", @"earth", @"weakTo", [UIColor lightGrayColor], @"color", nil],
+			[NSDictionary dictionaryWithObjectsAndKeys:[NSNumber numberWithInteger: RQElementalTypeFire], @"type", [UIColor redColor], @"color", nil],
+			[NSDictionary dictionaryWithObjectsAndKeys:[NSNumber numberWithInteger: RQElementalTypeWater], @"type", [UIColor blueColor], @"color", nil],
+			[NSDictionary dictionaryWithObjectsAndKeys:[NSNumber numberWithInteger: RQElementalTypeEarth], @"type", [UIColor brownColor], @"color", nil],
+			[NSDictionary dictionaryWithObjectsAndKeys:[NSNumber numberWithInteger: RQElementalTypeAir], @"type", [UIColor lightGrayColor], @"color", nil],
 			nil];
+}
+
+- (RQElementalType)weakToType
+{
+	RQElementalType answer = 0;
+		switch ([self.typeAsNumber integerValue]) {
+			case RQElementalTypeFire:
+				answer = RQElementalTypeWater;
+				break;
+			case RQElementalTypeWater:
+				answer = RQElementalTypeAir;
+				break;
+			case RQElementalTypeEarth:
+				answer = RQElementalTypeFire;
+				break;
+			case RQElementalTypeAir:
+				answer = RQElementalTypeEarth;
+				break;
+		}
+	return answer;
+}
+
+- (RQElementalType)strongToType
+{
+	RQElementalType answer = 0;
+	switch ([self.typeAsNumber integerValue]) {
+		case RQElementalTypeWater:
+			answer = RQElementalTypeFire;
+			break;
+		case RQElementalTypeAir:
+			answer = RQElementalTypeWater;
+			break;
+		case RQElementalTypeFire:
+			answer = RQElementalTypeEarth;
+			break;
+		case RQElementalTypeEarth:
+			answer = RQElementalTypeAir;
+			break;
+	}
+	return answer;
 }
 
 @end